+++ Changelog.md
@@ -2,6 +2,7 @@
 ## Alpha 0.0.5
 In development
 * Removed indentation from ModItems file to be reimplemented at a later date
+* A creative menu tab is now included to contain all modded items
 
 ## Alpha 0.0.4
 Released on 14th March 2018